Configures the test image with a custom DNS server.
This change allows users to set a custom DNS server
in the config.json of the test device. The
DNS server is only set if the environment variable
DNS_SERVER is set to a non-zero value.

This is needed to test with a balenaCloud instance
whose address can only be resolved using a custom
name server. This is a common setup for balenaMachine
deployments in air-gapped networks.

Change-type: patch
Signed-off-by: Carlo Miguel F. Cruz <carloc@balena.io>