qemu: test for kvm device node
Before running qemu, test for the kvm device node existing and being
read/write for the current user.

This handles an edge case where the CPU supports hardware virtualization
extensions, but the KVM device node doesn't exist or the user can't
access it, such as from inside a container, or when disabled in the
kernel's configuration.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>